requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.12
Flask-Migrate==2.5.2
psycopg2-binary==2.8.4
python-dotenv==0.11.0
firebase-admin==4.1.0
graphene==2.1.8
flask-graphql==2.0.1
